+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- `experimentalSetExplicitDimensions` prop for `Image` and `ImageList`
+
 ## [0.15.0] - 2022-09-08
 
 ### Added